remove existence check

+++ storage/elasticsearch/src/main/scala/org/apache/predictionio/data/storage/elasticsearch/ESSequences.scala
@@ -54,7 +54,7 @@ class ESSequences(client: ESClient, config: StorageClientConfig, index: String) 
     restClient.close()
   }
 
-  def genNext(name: String, restClient: RestClient): Int = {
+  def genNext(name: String, restClient: RestClient): Long = {
     try {
       val entity = new NStringEntity(write("n" -> name), ContentType.APPLICATION_JSON)
       val response = restClient.performRequest(
@@ -66,9 +66,9 @@ class ESSequences(client: ESClient, config: StorageClientConfig, index: String) 
       val result = (jsonResponse \ "result").extract[String]
       result match {
         case "created" =>
-          (jsonResponse \ "_version").extract[Int]
+          (jsonResponse \ "_version").extract[Long]
         case "updated" =>
-          (jsonResponse \ "_version").extract[Int]
+          (jsonResponse \ "_version").extract[Long]
         case _ =>
           throw new IllegalStateException(s"[$result] Failed to update $index/$estype/$name")
       }
